pschart, psfd: avoid integer overflows in averaging samples.
pschart calculates the mean value of each strip in order to
automatically position it on the page.  psfd calculates the mean of
the decimated samples, which are each calculated by averaging a
(potentially large) number of input samples.

These calculations are performed on WFDB_Sample values (which may be
as large as 32 bits, for format 32), and the sum was stored as a long,
so it could easily overflow if there were a large number of samples
(nsamp / nisamp / decf) or if the signals were high resolution.

Instead, avoid this issue by calculating the sums as doubles.

+++ app/pschart.c
@@ -542,7 +542,7 @@ int *uncal;		/* if non-zero, signal is uncalibrated */
 
 /* Arrays indexed by signal # (allocated by process(), used by printstrip()) */
 int *accept, *buflen, *v, *vbase, **vbuf, *vmax, *vmin;
-long *vsum;
+double *vsum;
 
 /* Derived parameters */
 double dpmm;		/* pixels per millimeter */
@@ -659,7 +659,7 @@ FILE *cfile;
 		    (vbase = realloc(vbase, nosig * sizeof(int))) == NULL ||
 		    (vmax = realloc(vmax, nosig * sizeof(int))) == NULL ||
 		    (vmin = realloc(vmin, nosig * sizeof(int))) == NULL ||
-		    (vsum = realloc(vsum, nosig * sizeof(long))) == NULL ||
+		    (vsum = realloc(vsum, nosig * sizeof(double))) == NULL ||
 		    (vbuf = realloc(vbuf, nosig * sizeof(int *))) == NULL) {
 		    (void)fprintf(stderr, "%s: insufficient memory\n", pname);
 		    exit(2);
